refactor: make interface and impl in same file
